Build with Go 1.17 for native Apple M1 support
This change updates all GitHub workflows from Go 1.14 to Go 1.17, in
particular to start building native darwin/arm64 binaries for Apple's
newest machines. Terraform doesn't account for Rosetta 2 emulation when
it selects providers, so without a native build of a given provider the
workaround is to run a darwin/amd64 build of Terraform and use amd64
provider binaries across the board.

As far as I can tell, the current GoReleaser matrix should build for
this OS + arch combination once a new enough Go version is in place.